Geoserver REST API DataStore

Geoserver REST API DataStore

Geoserver REST API DataStore

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer dataStore

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geosapi::GSRESTResource -> geosapi::GSAbstractStore -> GSAbstractDataStore

Public fields

connectionParameters

the list of connection parameters

Methods

Public methods

GSAbstractDataStore$new()

GSAbstractDataStore$decode()

GSAbstractDataStore$setConnectionParameters()

GSAbstractDataStore$addConnectionParameter()

GSAbstractDataStore$setConnectionParameter()

GSAbstractDataStore$delConnectionParameter()

GSAbstractDataStore$clone()

Inherited methods

geosapi::GSRESTResource$encode()

geosapi::GSAbstractStore$setDescription()

geosapi::GSAbstractStore$setEnabled()

geosapi::GSAbstractStore$setType()

Method new()

initializes an abstract data store

Usage

GSAbstractDataStore$new(
  xml = NULL,
  type = NULL,
  name = NULL,
  description = "",
  enabled = TRUE,
  connectionParameters
)

Arguments

xml

an object of class XMLInternalNode-class to create object from XML

type

the type of coverage store

name

coverage store name

description

coverage store description

enabled

whether the store should be enabled or not. Default is TRUE

connectionParameters

the list of connection parameters

Method decode()

Decodes a data store from XML

Usage

GSAbstractDataStore$decode(xml)

Arguments

xml

an object of class XMLInternalNode-class

Returns

an object of class GSAbstractDataStore

Method setConnectionParameters()

Set list connection parameters. The argument should be an object of class GSRESTEntrySet giving a list of key/value parameter entries.

Usage

GSAbstractDataStore$setConnectionParameters(parameters)

Arguments

parameters

an object of class GSRESTEntrySet

Method addConnectionParameter()

Adds a connection parameter

Usage

GSAbstractDataStore$addConnectionParameter(key, value)

Arguments

key

connection parameter key

value

connection parameter value

Returns

TRUE if added, FALSE otherwise

Method setConnectionParameter()

Sets a connection parameter

Usage

GSAbstractDataStore$setConnectionParameter(key, value)

Arguments

key

connection parameter key

value

connection parameter value

Method delConnectionParameter()

Removes a connection parameter

Usage

GSAbstractDataStore$delConnectionParameter(key)

Arguments

key

connection parameter key

value

connection parameter value

Returns

TRUE if removed, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

GSAbstractDataStore$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.