Geoserver REST API GSCoverageView

Geoserver REST API GSCoverageView

Geoserver REST API GSCoverageView

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer coverage view

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSRESTResource -> GSCoverageView

Public fields

name

name

envelopeCompositionType

envelope composition type

selectedResolution

selected resolution

selectedResolutionIndex

selected resolution index

coverageBands

coverage bands

Methods

Public methods

GSCoverageView$new()

GSCoverageView$decode()

GSCoverageView$setName()

GSCoverageView$setEnvelopeCompositionType()

GSCoverageView$setSelectedResolution()

GSCoverageView$setSelectedResolutionIndex()

GSCoverageView$addBand()

GSCoverageView$delBand()

GSCoverageView$clone()

Inherited methods

geosapi::GSRESTResource$encode()

Method new()

Initializes an object of class GSCoverageView

Usage

GSCoverageView$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method decode()

Decodes from XML

Usage

GSCoverageView$decode(xml)

Arguments

xml

object of class XMLInternalNode-class

Method setName()

Set name

Usage

GSCoverageView$setName(name)

Arguments

name

name

Method setEnvelopeCompositionType()

Sets the envelope composition type. Type of Envelope Composition, used to expose the bounding box of the CoverageView, either 'UNION' or 'INTERSECTION'.

Usage

GSCoverageView$setEnvelopeCompositionType(envelopeCompositionType)

Arguments

envelopeCompositionType

envelope composition type

Method setSelectedResolution()

Set selected resolution

Usage

GSCoverageView$setSelectedResolution(selectedResolution)

Arguments

selectedResolution

selected resolution

Method setSelectedResolutionIndex()

Set selected resolution index

Usage

GSCoverageView$setSelectedResolutionIndex(selectedResolutionIndex)

Arguments

selectedResolutionIndex

selected resolution index

Method addBand()

Adds band

Usage

GSCoverageView$addBand(band)

Arguments

band

object of class GSCoverageBand

Returns

TRUE if added, FALSE otherwise

Method delBand()

Deletes band

Usage

GSCoverageView$delBand(band)

Arguments

band

object of class GSCoverageBand

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

GSCoverageView$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  GSCoverageView$new()
#> <GSCoverageView>
#>   Inherits from: <GSRESTResource>
#>   Public:
#>     addBand: function (band) 
#>     clone: function (deep = FALSE) 
#>     coverageBands: list
#>     decode: function (xml) 
#>     delBand: function (band) 
#>     encode: function () 
#>     envelopeCompositionType: NULL
#>     initialize: function (xml = NULL) 
#>     name: NA
#>     rootName: coverageView
#>     selectedResolution: NULL
#>     selectedResolutionIndex: NULL
#>     setEnvelopeCompositionType: function (envelopeCompositionType) 
#>     setName: function (name) 
#>     setSelectedResolution: function (selectedResolution) 
#>     setSelectedResolutionIndex: function (selectedResolutionIndex)