A GeoServer feature type

This class models a GeoServer feature type. This class is to be used for manipulating representations of vector data with GeoServer.

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer feature type

Details

Geoserver REST API Resource

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geosapi::GSRESTResource -> geosapi::GSResource -> GSFeatureType

Public fields

cqlFilter

CQL filter

Methods

Public methods

GSFeatureType$new()

GSFeatureType$decode()

GSFeatureType$setCqlFilter()

GSFeatureType$setVirtualTable()

GSFeatureType$delVirtualTable()

GSFeatureType$clone()

Inherited methods

geosapi::GSRESTResource$encode()

geosapi::GSRESTResource$getClassName()

geosapi::GSRESTResource$print()

geosapi::GSResource$addKeyword()

geosapi::GSResource$addMetadataLink()

geosapi::GSResource$delKeyword()

geosapi::GSResource$delMetadata()

geosapi::GSResource$deleteMetadataLink()

geosapi::GSResource$setAbstract()

geosapi::GSResource$setDescription()

geosapi::GSResource$setEnabled()

geosapi::GSResource$setKeywords()

geosapi::GSResource$setLatLonBoundingBox()

geosapi::GSResource$setMetadata()

geosapi::GSResource$setMetadataDimension()

geosapi::GSResource$setMetadataLinks()

geosapi::GSResource$setName()

geosapi::GSResource$setNativeBoundingBox()

geosapi::GSResource$setNativeCRS()

geosapi::GSResource$setNativeName()

geosapi::GSResource$setProjectionPolicy()

geosapi::GSResource$setSrs()

geosapi::GSResource$setTitle()

Method new()

Initializes an object of class GSFeatureType

Usage

GSFeatureType$new(xml = NULL)

Arguments

xml

object of class xml_node-class

Method decode()

Decodes from XML

Usage

GSFeatureType$decode(xml)

Arguments

xml

object of class xml_node-class

Method setCqlFilter()

Set CQL filter

Usage

GSFeatureType$setCqlFilter(cqlFilter)

Arguments

cqlFilter

CQL filter

Method setVirtualTable()

Set virtual table

Usage

GSFeatureType$setVirtualTable(vt)

Arguments

vt

object of class GSVirtualTable

Returns

TRUE if set/added, FALSE otherwise

Method delVirtualTable()

Deletes virtual table

Usage

GSFeatureType$delVirtualTable()

Arguments

vt

object of class GSVirtualTable

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

GSFeatureType$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  ft <- GSFeatureType$new()