Geoserver REST API GeoPackageDataStore

Geoserver REST API GeoPackageDataStore

Geoserver REST API GeoPackageDataStore

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer GeoPackage dataStore

Methods inherited from GSAbstractDBDataStore

setDatabaseType(dbtype)

Sets the database type, here "geopkg

setNamespace(namespace)

Sets the datastore namespace

setHost(host)

Sets the database host

setPort(port)

Set the database port

setDatabase(database)

Set the database name

setSchema(schema)

Set the database schema

setUser(user)

Set the database username

setPassword(password)

Set the database password

setJndiReferenceName(jndiReferenceName)

Set a JNDI reference name

setExposePrimaryKeys(exposePrimaryKeys)

Set TRUE if primary keys have to be exposed to datastore, FALSE otherwise.

setMaxConnections(maxConnections)

Set the maximum number of connections. Default is set to 10.

setMinConnections(minConnections)

Set the minimum number of connections. Default is set to 1.

setFetchSize(fetchSize)

Set the fetch size. Default is set to 10.

setConnectionTimeout(seconds)

Set the connection timeout. Default is set to 20s.

setValidateConnections(validateConnections)

Set TRUE if connections have to be validated, FALSE otherwise.

setPrimaryKeyMetadataTable(primaryKeyMetadataTable)

Set the name of the primaryKey metadata table

setLooseBBox(looseBBox)

Set losse bbox parameter.

setPreparedStatements(preparedStatements)

Set prepared statements

setMaxOpenPreparedStatements(maxOpenPreparedStatements)

Set maximum open prepared statements

setEstimatedExtends(estimatedExtends)

Set estimatedExtend parameter

setDefautConnectionParameters()

Set default connection parameters

Methods

new(xml, name, description, enabled, database)

Instantiates a GSGeoPackageDataStore object

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geosapi::GSRESTResource -> geosapi::GSAbstractStore -> geosapi::GSAbstractDataStore -> geosapi::GSAbstractDBDataStore -> GSGeoPackageDataStore

Methods

Public methods

GSGeoPackageDataStore$new()

GSGeoPackageDataStore$clone()

Inherited methods

geosapi::GSRESTResource$encode()

geosapi::GSAbstractStore$setDescription()

geosapi::GSAbstractStore$setEnabled()

geosapi::GSAbstractStore$setType()

geosapi::GSAbstractDataStore$addConnectionParameter()

geosapi::GSAbstractDataStore$decode()

geosapi::GSAbstractDataStore$delConnectionParameter()

geosapi::GSAbstractDataStore$setConnectionParameter()

geosapi::GSAbstractDataStore$setConnectionParameters()

geosapi::GSAbstractDBDataStore$setConnectionTimeout()

geosapi::GSAbstractDBDataStore$setDatabase()

geosapi::GSAbstractDBDataStore$setDatabaseType()

geosapi::GSAbstractDBDataStore$setDefautConnectionParameters()

geosapi::GSAbstractDBDataStore$setEstimatedExtends()

geosapi::GSAbstractDBDataStore$setExposePrimaryKeys()

geosapi::GSAbstractDBDataStore$setFetchSize()

geosapi::GSAbstractDBDataStore$setHost()

geosapi::GSAbstractDBDataStore$setJndiReferenceName()

geosapi::GSAbstractDBDataStore$setLooseBBox()

geosapi::GSAbstractDBDataStore$setMaxConnections()

geosapi::GSAbstractDBDataStore$setMaxOpenPreparedStatements()

geosapi::GSAbstractDBDataStore$setMinConnections()

geosapi::GSAbstractDBDataStore$setNamespace()

geosapi::GSAbstractDBDataStore$setPassword()

geosapi::GSAbstractDBDataStore$setPort()

geosapi::GSAbstractDBDataStore$setPreparedStatements()

geosapi::GSAbstractDBDataStore$setPrimaryKeyMetadataTable()

geosapi::GSAbstractDBDataStore$setSchema()

geosapi::GSAbstractDBDataStore$setUser()

geosapi::GSAbstractDBDataStore$setValidateConnections()

Method new()

initializes an GeoPackage data store

Usage

GSGeoPackageDataStore$new(
  xml = NULL,
  name = NULL,
  description = "",
  enabled = TRUE,
  database = NULL
)

Arguments

xml

an object of class XMLInternalNode-class to create object from XML

name

coverage store name

description

coverage store description

enabled

whether the store should be enabled or not. Default is TRUE

database

database

Method clone()

The objects of this class are cloneable with this method.

Usage

GSGeoPackageDataStore$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

ds <- GSGeoPackageDataStore$new(
 name = "ds", description = "des", 
 enabled = TRUE, database = NULL
)