A GeoServer layer resource

This class models a GeoServer layer. This class is to be used for published resource (feature type or coverage).

This class models a GeoServer style.

Format

R6Class object.

R6Class object.

Value

Object of R6Class for modelling a GeoServer layer

Object of R6Class for modelling a GeoServer style

Details

Geoserver REST API Resource

Geoserver REST API Style

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSRESTResource -> GSLayer

Public fields

full

full

name

name

path

path

defaultStyle

default style

styles

styles

enabled

enabled

queryable

queryable

advertised

advertised

Methods

Public methods

GSLayer$new()

GSLayer$decode()

GSLayer$setName()

GSLayer$setPath()

GSLayer$setEnabled()

GSLayer$setQueryable()

GSLayer$setAdvertised()

GSLayer$setDefaultStyle()

GSLayer$setStyles()

GSLayer$addStyle()

GSLayer$delStyle()

GSLayer$clone()

Inherited methods

geosapi::GSRESTResource$encode()

Method new()

Initializes an object of class GSLayer

Usage

GSLayer$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method decode()

Decodes from XML

Usage

GSLayer$decode(xml)

Arguments

xml

object of class XMLInternalNode-class

Method setName()

Set name

Usage

GSLayer$setName(name)

Arguments

name

name

Method setPath()

Set path

Usage

GSLayer$setPath(path)

Arguments

path

path

Method setEnabled()

Set enabled

Usage

GSLayer$setEnabled(enabled)

Arguments

enabled

enabled

Method setQueryable()

Set queryable

Usage

GSLayer$setQueryable(queryable)

Arguments

queryable

queryable

Method setAdvertised()

Set advertised

Usage

GSLayer$setAdvertised(advertised)

Arguments

advertised

advertised

Method setDefaultStyle()

Set default style

Usage

GSLayer$setDefaultStyle(style)

Arguments

style

object o class GSStyle or character

Method setStyles()

Set styles

Usage

GSLayer$setStyles(styles)

Arguments

styles

styles

Method addStyle()

Adds style

Usage

GSLayer$addStyle(style)

Arguments

style

style, object o class GSStyle or character

Returns

TRUE if added, FALSE otherwise

Method delStyle()

Deletes style

Usage

GSLayer$delStyle(style)

Arguments

style

style, object o class GSStyle or character

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

GSLayer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super class

geosapi::GSRESTResource -> GSStyle

Public fields

full

full

name

name

filename

filename

Methods

Public methods

GSStyle$new()

GSStyle$decode()

GSStyle$setName()

GSStyle$setFilename()

GSStyle$clone()

Inherited methods

geosapi::GSRESTResource$encode()

Method new()

Initializes a GSStyle

Usage

GSStyle$new(xml = NULL, name = NULL, filename = NULL)

Arguments

xml

an object of class XMLInternalNode-class

name

name

filename

filename

Method decode()

Decodes from XML

Usage

GSStyle$decode(xml)

Arguments

xml

an object of class XMLInternalNode-class

Method setName()

set name

Usage

GSStyle$setName(name)

Arguments

name

name

Method setFilename()

Set filename

Usage

GSStyle$setFilename(filename)

Arguments

filename

filename

Method clone()

The objects of this class are cloneable with this method.

Usage

GSStyle$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  lyr <- GSLayer$new()

  lyr <- GSStyle$new()