A GeoServer layergroup resource

This class models a GeoServer layer group. This class is to be used for clustering layers into a group.

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer layergroup

Details

Geoserver REST API LayerGroup

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSRESTResource -> GSLayerGroup

Public fields

full

full

name

name

mode

mode

title

title

abstractTxt

abstract

workspace

workspace

publishables

publishables

styles

styles

metadataLinks

metadata links

bounds

bounds

Methods

Public methods

GSLayerGroup$new()

GSLayerGroup$decode()

GSLayerGroup$setName()

GSLayerGroup$setMode()

GSLayerGroup$setTitle()

GSLayerGroup$setAbstract()

GSLayerGroup$setWorkspace()

GSLayerGroup$addLayer()

GSLayerGroup$addLayerGroup()

GSLayerGroup$addPublishable()

GSLayerGroup$setStyles()

GSLayerGroup$addStyle()

GSLayerGroup$setMetadataLinks()

GSLayerGroup$addMetadataLink()

GSLayerGroup$deleteMetadataLink()

GSLayerGroup$setBounds()

GSLayerGroup$clone()

Inherited methods

geosapi::GSRESTResource$encode()

geosapi::GSRESTResource$getClassName()

geosapi::GSRESTResource$print()

Method new()

Initializes an object of class GSLayerGroup

Usage

GSLayerGroup$new(xml = NULL)

Arguments

xml

object of class xml_node-class

Method decode()

Decodes from XML

Usage

GSLayerGroup$decode(xml)

Arguments

xml

object of class xml_node-class

Method setName()

Set name

Usage

GSLayerGroup$setName(name)

Arguments

name

name

Method setMode()

Set mode

Usage

GSLayerGroup$setMode(mode)

Arguments

mode

a mode value among "SINGLE", "NAMED", "CONTAINER", "EO"

Method setTitle()

Set title

Usage

GSLayerGroup$setTitle(title)

Arguments

title

title

Method setAbstract()

Set abstract

Usage

GSLayerGroup$setAbstract(abstract)

Arguments

abstract

abstract

Method setWorkspace()

Set workspace

Usage

GSLayerGroup$setWorkspace(workspace)

Arguments

workspace

workspace name, object of class GSWorkspace or character

Method addLayer()

Adds layer

Usage

GSLayerGroup$addLayer(layer, style)

Arguments

layer

layer name

style

style name

Method addLayerGroup()

Adds layer group

Usage

GSLayerGroup$addLayerGroup(layerGroup)

Arguments

layerGroup

layer group

Method addPublishable()

Adds publishable

Usage

GSLayerGroup$addPublishable(publishable)

Arguments

publishable

publishable

Returns

TRUE if added, FALSE otherwise

Method setStyles()

Set styles

Usage

GSLayerGroup$setStyles(styles)

Arguments

styles

styles

Method addStyle()

Adds a style

Usage

GSLayerGroup$addStyle(style)

Arguments

style

style

Returns

TRUE if added, FALSE otherwise

Method setMetadataLinks()

Set metadata links

Usage

GSLayerGroup$setMetadataLinks(metadataLinks)

Arguments

metadataLinks

metadata links

Method addMetadataLink()

Adds metadata link

Usage

GSLayerGroup$addMetadataLink(metadataLink)

Arguments

metadataLink

object of class GSMetadataLink

Returns

TRUE if added, FALSE otherwise

Method deleteMetadataLink()

Deletes metadata link

Usage

GSLayerGroup$deleteMetadataLink(metadataLink)

Arguments

metadataLink

object of class GSMetadataLink

Returns

TRUE if deleted, FALSE otherwise

Method setBounds()

Set bounds

Usage

GSLayerGroup$setBounds(minx, miny, maxx, maxy, bbox = NULL, crs)

Arguments

minx

minx

miny

miny

maxx

maxx

maxy

maxy

bbox

bbox

crs

crs

Method clone()

The objects of this class are cloneable with this method.

Usage

GSLayerGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  lyr <- GSLayerGroup$new()