Geoserver REST API Namespace Manager

Geoserver REST API Namespace Manager

Geoserver REST API Namespace Manager

Format

R6Class object.

Value

Object of R6Class with methods for managing the namespaces of a GeoServer instance.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSManager -> GSNamespaceManager

Methods

Public methods

GSNamespaceManager$getNamespaces()

GSNamespaceManager$getNamespaceNames()

GSNamespaceManager$getNamespace()

GSNamespaceManager$createNamespace()

GSNamespaceManager$updateNamespace()

GSNamespaceManager$deleteNamespace()

GSNamespaceManager$clone()

Inherited methods

geosapi::GSManager$ERROR()

geosapi::GSManager$INFO()

geosapi::GSManager$WARN()

geosapi::GSManager$connect()

geosapi::GSManager$getClassName()

geosapi::GSManager$getCoverageStoreManager()

geosapi::GSManager$getDataStoreManager()

geosapi::GSManager$getNamespaceManager()

geosapi::GSManager$getServiceManager()

geosapi::GSManager$getStyleManager()

geosapi::GSManager$getSystemStatus()

geosapi::GSManager$getUrl()

geosapi::GSManager$getWorkspaceManager()

geosapi::GSManager$initialize()

geosapi::GSManager$logger()

geosapi::GSManager$monitor()

geosapi::GSManager$reload()

Method getNamespaces()

Get the list of available namespace. Re

Usage

GSNamespaceManager$getNamespaces()

Returns

an object of class list containing items of class GSNamespace

Method getNamespaceNames()

Get the list of available namespace names.

Usage

GSNamespaceManager$getNamespaceNames()

Returns

a vector of class character

Method getNamespace()

Get a GSNamespace object given a namespace name.

Usage

GSNamespaceManager$getNamespace(ns)

Arguments

ns

namespace

Returns

an object of class GSNamespace

Method createNamespace()

Creates a GeoServer namespace given a prefix, and an optional URI.

Usage

GSNamespaceManager$createNamespace(prefix, uri)

Arguments

prefix

prefix

uri

uri

Returns

TRUE if the namespace has been successfully created, FALSE otherwise

Method updateNamespace()

Updates a GeoServer namespace given a prefix, and an optional URI.

Usage

GSNamespaceManager$updateNamespace(prefix, uri)

Arguments

prefix

prefix

uri

uri

Returns

TRUE if the namespace has been successfully updated, FALSE otherwise

Method deleteNamespace()

Deletes a GeoServer namespace given a name.

Usage

GSNamespaceManager$deleteNamespace(name, recurse = FALSE)

Arguments

name

name

recurse

recurse

Returns

TRUE if the namespace has been successfully deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

GSNamespaceManager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \dontrun{
   GSNamespaceManager$new("http://localhost:8080/geoserver", "admin", "geoserver")
#> Error in curl::curl_fetch_memory(url, handle = handle): Couldn't connect to server [localhost]: Failed to connect to localhost port 8080 after 0 ms: Couldn't connect to server
# }