Geoserver REST API OracleNGDataStore

Geoserver REST API OracleNGDataStore

Geoserver REST API OracleNGDataStore

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer OracleNG dataStore

Methods inherited from GSAbstractDBDataStore

setDatabaseType(dbtype)

Sets the database type, here "OracleNG"

setNamespace(namespace)

Sets the datastore namespace

setHost(host)

Sets the database host

setPort(port)

Set the database port

setDatabase(database)

Set the database name

setSchema(schema)

Set the database schema

setUser(user)

Set the database username

setPassword(password)

Set the database password

setJndiReferenceName(jndiReferenceName)

Set a JNDI reference name

setExposePrimaryKeys(exposePrimaryKeys)

Set TRUE if primary keys have to be exposed to datastore, FALSE otherwise.

setMaxConnections(maxConnections)

Set the maximum number of connections. Default is set to 10.

setMinConnections(minConnections)

Set the minimum number of connections. Default is set to 1.

setFetchSize(fetchSize)

Set the fetch size. Default is set to 10.

setConnectionTimeout(seconds)

Set the connection timeout. Default is set to 20s.

setValidateConnections(validateConnections)

Set TRUE if connections have to be validated, FALSE otherwise.

setPrimaryKeyMetadataTable(primaryKeyMetadataTable)

Set the name of the primaryKey metadata table

setLooseBBox(looseBBox)

Set losse bbox parameter.

setPreparedStatements(preparedStatements)

Set prepared statements

setMaxOpenPreparedStatements(maxOpenPreparedStatements)

Set maximum open prepared statements

setEstimatedExtends(estimatedExtends)

Set estimatedExtend parameter

setDefautConnectionParameters()

Set default connection parameters

Methods

new(xml, name, description, enabled)

Instantiates a GSOracleNGDataStore object

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geosapi::GSRESTResource -> geosapi::GSAbstractStore -> geosapi::GSAbstractDataStore -> geosapi::GSAbstractDBDataStore -> GSOracleNGDataStore

Methods

Public methods

GSOracleNGDataStore$new()

GSOracleNGDataStore$clone()

Inherited methods

geosapi::GSRESTResource$encode()

geosapi::GSAbstractStore$setDescription()

geosapi::GSAbstractStore$setEnabled()

geosapi::GSAbstractStore$setType()

geosapi::GSAbstractDataStore$addConnectionParameter()

geosapi::GSAbstractDataStore$decode()

geosapi::GSAbstractDataStore$delConnectionParameter()

geosapi::GSAbstractDataStore$setConnectionParameter()

geosapi::GSAbstractDataStore$setConnectionParameters()

geosapi::GSAbstractDBDataStore$setConnectionTimeout()

geosapi::GSAbstractDBDataStore$setDatabase()

geosapi::GSAbstractDBDataStore$setDatabaseType()

geosapi::GSAbstractDBDataStore$setDefautConnectionParameters()

geosapi::GSAbstractDBDataStore$setEstimatedExtends()

geosapi::GSAbstractDBDataStore$setExposePrimaryKeys()

geosapi::GSAbstractDBDataStore$setFetchSize()

geosapi::GSAbstractDBDataStore$setHost()

geosapi::GSAbstractDBDataStore$setJndiReferenceName()

geosapi::GSAbstractDBDataStore$setLooseBBox()

geosapi::GSAbstractDBDataStore$setMaxConnections()

geosapi::GSAbstractDBDataStore$setMaxOpenPreparedStatements()

geosapi::GSAbstractDBDataStore$setMinConnections()

geosapi::GSAbstractDBDataStore$setNamespace()

geosapi::GSAbstractDBDataStore$setPassword()

geosapi::GSAbstractDBDataStore$setPort()

geosapi::GSAbstractDBDataStore$setPreparedStatements()

geosapi::GSAbstractDBDataStore$setPrimaryKeyMetadataTable()

geosapi::GSAbstractDBDataStore$setSchema()

geosapi::GSAbstractDBDataStore$setUser()

geosapi::GSAbstractDBDataStore$setValidateConnections()

Method new()

initializes an Oracle NG data store

Usage

GSOracleNGDataStore$new(
  xml = NULL,
  name = NULL,
  description = "",
  enabled = TRUE
)

Arguments

xml

an object of class XMLInternalNode-class to create object from XML

name

coverage store name

description

coverage store description

enabled

whether the store should be enabled or not. Default is TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

GSOracleNGDataStore$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  GSOracleNGDataStore$new(name = "ds", description = "des", enabled = TRUE)
#> <GSOracleNGDataStore>
#>   Inherits from: <GSAbstractDBDataStore>
#>   Public:
#>     addConnectionParameter: function (key, value) 
#>     clone: function (deep = FALSE) 
#>     connectionParameters: GSRESTEntrySet, GSRESTResource, R6
#>     decode: function (xml) 
#>     delConnectionParameter: function (key) 
#>     description: des
#>     enabled: TRUE
#>     encode: function () 
#>     full: TRUE
#>     initialize: function (xml = NULL, name = NULL, description = "", enabled = TRUE) 
#>     name: ds
#>     rootName: dataStore
#>     setConnectionParameter: function (key, value) 
#>     setConnectionParameters: function (parameters) 
#>     setConnectionTimeout: function (seconds = 20) 
#>     setDatabase: function (database) 
#>     setDatabaseType: function (dbtype) 
#>     setDefautConnectionParameters: function () 
#>     setDescription: function (description) 
#>     setEnabled: function (enabled) 
#>     setEstimatedExtends: function (estimatedExtends = FALSE) 
#>     setExposePrimaryKeys: function (exposePrimaryKeys) 
#>     setFetchSize: function (fetchSize = 1000) 
#>     setHost: function (host) 
#>     setJndiReferenceName: function (jndiReferenceName) 
#>     setLooseBBox: function (looseBBox = TRUE) 
#>     setMaxConnections: function (maxConnections = 10) 
#>     setMaxOpenPreparedStatements: function (maxOpenPreparedStatements = 50) 
#>     setMinConnections: function (minConnections = 1) 
#>     setNamespace: function (namespace) 
#>     setPassword: function (password) 
#>     setPort: function (port) 
#>     setPreparedStatements: function (preparedStatements = FALSE) 
#>     setPrimaryKeyMetadataTable: function (primaryKeyMetadataTable) 
#>     setSchema: function (schema) 
#>     setType: function (type) 
#>     setUser: function (user) 
#>     setValidateConnections: function (validateConnections) 
#>     type: Oracle NG
#>     workspace: NULL
#>   Private:
#>     DBTYPE: oracle
#>     STORE_TYPE: dataStore
#>     TYPE: Oracle NG
#>     store_type: dataStore