A GeoServer abstract resource

This class models an abstract GeoServer resource. This class is used internally for modelling instances of class GSFeatureType or GSCoverage

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer resource

Details

Geoserver REST API Resource

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSRESTResource -> GSResource

Public fields

full

full

name

resource name

nativeName

resource native name

title

resource title

description

resource description

abstract

resource abstract

keywords

resource keywords

metadataLinks

resource metadata links

nativeCRS

resource native CRS

srs

resource srs

nativeBoundingBox

resource lat/lon native bounding box

latLonBoundingBox

resource lat/lon bounding box

projectionPolicy

resource projection policy

enabled

enabled

metadata

metadata

Methods

Public methods

GSResource$new()

GSResource$decode()

GSResource$setEnabled()

GSResource$setName()

GSResource$setNativeName()

GSResource$setTitle()

GSResource$setDescription()

GSResource$setAbstract()

GSResource$setKeywords()

GSResource$addKeyword()

GSResource$delKeyword()

GSResource$setMetadataLinks()

GSResource$addMetadataLink()

GSResource$deleteMetadataLink()

GSResource$setProjectionPolicy()

GSResource$setSrs()

GSResource$setNativeCRS()

GSResource$setLatLonBoundingBox()

GSResource$setNativeBoundingBox()

GSResource$setMetadata()

GSResource$delMetadata()

GSResource$setMetadataDimension()

GSResource$clone()

Inherited methods

geosapi::GSRESTResource$encode()

geosapi::GSRESTResource$getClassName()

geosapi::GSRESTResource$print()

Method new()

Initializes a GSResource

Usage

GSResource$new(rootName = NULL, xml = NULL)

Arguments

rootName

root name

xml

object of class xml_node-class

Method decode()

Decodes from XML

Usage

GSResource$decode(xml)

Arguments

xml

object of class xml_node-class

Method setEnabled()

Set enabled

Usage

GSResource$setEnabled(enabled)

Arguments

enabled

enabled

Method setName()

Set name

Usage

GSResource$setName(name)

Arguments

name

name

Method setNativeName()

Set native name

Usage

GSResource$setNativeName(nativeName)

Arguments

nativeName

native name

Method setTitle()

Set title

Usage

GSResource$setTitle(title)

Arguments

title

title

Method setDescription()

Set description

Usage

GSResource$setDescription(description)

Arguments

description

description

Method setAbstract()

Set abstract

Usage

GSResource$setAbstract(abstract)

Arguments

abstract

abstract

Method setKeywords()

Set keyword(s)

Usage

GSResource$setKeywords(keywords)

Arguments

keywords

keywords

Method addKeyword()

Adds keyword

Usage

GSResource$addKeyword(keyword)

Arguments

keyword

keyword

Returns

TRUE if added, FALSE otherwise

Method delKeyword()

Deletes keyword

Usage

GSResource$delKeyword(keyword)

Arguments

keyword

keyword

Returns

TRUE if deleted, FALSE otherwise

Method setMetadataLinks()

Set metadata links

Usage

GSResource$setMetadataLinks(metadataLinks)

Arguments

metadataLinks

metadata links

Method addMetadataLink()

Adds metadata link

Usage

GSResource$addMetadataLink(metadataLink)

Arguments

metadataLink

object of class GSMetadataLink

Returns

TRUE if added, FALSE otherwise

Method deleteMetadataLink()

Deletes metadata link

Usage

GSResource$deleteMetadataLink(metadataLink)

Arguments

metadataLink

object of class GSMetadataLink

Returns

TRUE if deleted, FALSE otherwise

Method setProjectionPolicy()

Set projection policy

Usage

GSResource$setProjectionPolicy(projectionPolicy)

Arguments

projectionPolicy

projection policy

Method setSrs()

Set SRS

Usage

GSResource$setSrs(srs)

Arguments

srs

srs

Method setNativeCRS()

Set native CRS

Usage

GSResource$setNativeCRS(nativeCRS)

Arguments

nativeCRS

native crs

Method setLatLonBoundingBox()

Set LatLon bounding box

Usage

GSResource$setLatLonBoundingBox(minx, miny, maxx, maxy, bbox = NULL, crs)

Arguments

minx

minx

miny

miny

maxx

maxx

maxy

maxy

bbox

bbox

crs

crs

Method setNativeBoundingBox()

Set native bounding box

Usage

GSResource$setNativeBoundingBox(minx, miny, maxx, maxy, bbox = NULL, crs)

Arguments

minx

minx

miny

miny

maxx

maxx

maxy

maxy

bbox

bbox

crs

crs

Method setMetadata()

Set metadata

Usage

GSResource$setMetadata(key, metadata)

Arguments

key

key

metadata

metadata

Returns

TRUE if added, FALSE otherwise

Method delMetadata()

Deletes metadata

Usage

GSResource$delMetadata(key)

Arguments

key

key

Returns

TRUE if deleted, FALSE otherwise

Method setMetadataDimension()

Set metadata dimension

Usage

GSResource$setMetadataDimension(key, dimension, custom = FALSE)

Arguments

key

key

dimension

dimension

custom

custom

Method clone()

The objects of this class are cloneable with this method.

Usage

GSResource$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

res <- GSResource$new(rootName = "featureType")