Geoserver REST API GSVirtualTable

Geoserver REST API GSVirtualTable

Geoserver REST API GSVirtualTable

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer virtual table

Methods

new(xml)

This method is used to instantiate a GSVirtualTable

decode(xml)

This method is used to decode a GSVirtualTable from XML

encode()

This method is used to encode a GSVirtualTable to XML

setName(name)

Sets the name of the virtual table

setSql(sql)

Sets the sql of the virtual table

setEscapeSql(escapeSql)

Sets the escapeSql. Default is FALSE

setKeyColumn(keyColumn)

Sets the keyColumn. Name of the column to be the primary key

setGeometry(vtg)

Sets the virtual table geometry

addParameter(parameter)

Adds a virtual table parameter

delParameter(parameter)

Removes a virtual table parameter.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSRESTResource -> GSVirtualTable

Public fields

name

name

sql

SQL statement

escapeSql

escape SQL?

keyColumn

key column

geometry

geometry

parameters

list of virtual parameters

Methods

Public methods

GSVirtualTable$new()

GSVirtualTable$decode()

GSVirtualTable$setName()

GSVirtualTable$setSql()

GSVirtualTable$setEscapeSql()

GSVirtualTable$setKeyColumn()

GSVirtualTable$setGeometry()

GSVirtualTable$addParameter()

GSVirtualTable$delParameter()

GSVirtualTable$clone()

Inherited methods

geosapi::GSRESTResource$encode()

Method new()

Initializes an object of class GSVirtualTable

Usage

GSVirtualTable$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method decode()

Decodes from XML

Usage

GSVirtualTable$decode(xml)

Arguments

xml

object of class XMLInternalNode-class

Method setName()

Set name

Usage

GSVirtualTable$setName(name)

Arguments

name

name

Method setSql()

Set SQL

Usage

GSVirtualTable$setSql(sql)

Arguments

sql

sql

Method setEscapeSql()

Set escape SQL

Usage

GSVirtualTable$setEscapeSql(escapeSql)

Arguments

escapeSql

escape SQL

Method setKeyColumn()

Set key column

Usage

GSVirtualTable$setKeyColumn(keyColumn)

Arguments

keyColumn

key column

Method setGeometry()

Set geometry

Usage

GSVirtualTable$setGeometry(vtg)

Arguments

vtg

object of class GSVirtualTableGeometry

Method addParameter()

Adds parameter

Usage

GSVirtualTable$addParameter(parameter)

Arguments

parameter

object of class GSVirtualTableParameter

Returns

TRUE if added, FALSE otherwise

Method delParameter()

Deletes parameter

Usage

GSVirtualTable$delParameter(parameter)

Arguments

parameter

object of class GSVirtualTableParameter

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

GSVirtualTable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

GSVirtualTable$new()
#> <GSVirtualTable>
#>   Inherits from: <GSRESTResource>
#>   Public:
#>     addParameter: function (parameter) 
#>     clone: function (deep = FALSE) 
#>     decode: function (xml) 
#>     delParameter: function (parameter) 
#>     encode: function () 
#>     escapeSql: FALSE
#>     geometry: NULL
#>     initialize: function (xml = NULL) 
#>     keyColumn: NULL
#>     name: NA
#>     parameters: list
#>     rootName: virtualTable
#>     setEscapeSql: function (escapeSql) 
#>     setGeometry: function (vtg) 
#>     setKeyColumn: function (keyColumn) 
#>     setName: function (name) 
#>     setSql: function (sql) 
#>     sql: NA