Geoserver REST API GSCoverageBand

Geoserver REST API GSCoverageBand

Geoserver REST API GSCoverageBand

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer coverage band

Methods

new(xml)

This method is used to instantiate a GSCoverageBand

decode(xml)

This method is used to decode a GSCoverageBand from XML

encode()

This method is used to encode a GSCoverageBand to XML

setDefinition(definition)

Sets the coverage band definition

setIndex(index)

Sets the coverage band index

setCompositionType

Sets the composition type. Only 'BAND_SELECT' is supported by GeoServer for now.

addInputBand(band)

Adds a input coverage band, object of class GSInputCoverageBand

delInputBand(band)

Removes a input coverage band, object of class GSInputCoverageBand

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSRESTResource -> GSCoverageBand

Public fields

inputCoverageBands

list of input coverage bands

definition

coverage band definition

index

coverage band index

compositionType

coverage band composition type

Methods

Public methods

GSCoverageBand$new()

GSCoverageBand$decode()

GSCoverageBand$setName()

GSCoverageBand$setDefinition()

GSCoverageBand$setIndex()

GSCoverageBand$setCompositionType()

GSCoverageBand$addInputBand()

GSCoverageBand$delInputBand()

GSCoverageBand$clone()

Inherited methods

geosapi::GSRESTResource$encode()

Method new()

Initalizes a GSCoverageBand

Usage

GSCoverageBand$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method decode()

Decodes from XML

Usage

GSCoverageBand$decode(xml)

Arguments

xml

object of class XMLInternalNode-class

Method setName()

Set name

Usage

GSCoverageBand$setName(name)

Arguments

name

name

Method setDefinition()

Set definition

Usage

GSCoverageBand$setDefinition(definition)

Arguments

definition

definition

Method setIndex()

Set index

Usage

GSCoverageBand$setIndex(index)

Arguments

index

index

Method setCompositionType()

Set composition type

Usage

GSCoverageBand$setCompositionType(compositionType)

Arguments

compositionType

composition type

Method addInputBand()

Adds an input band

Usage

GSCoverageBand$addInputBand(band)

Arguments

band

object of class GSInputCoverageBand

Returns

TRUE if added, FALSE otherwise

Method delInputBand()

Deletes an input band

Usage

GSCoverageBand$delInputBand(band)

Arguments

band

object of class GSInputCoverageBand

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

GSCoverageBand$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

GSCoverageBand$new()
#> <GSCoverageBand>
#>   Inherits from: <GSRESTResource>
#>   Public:
#>     addInputBand: function (band) 
#>     clone: function (deep = FALSE) 
#>     compositionType: BAND_SELECT
#>     decode: function (xml) 
#>     definition: NULL
#>     delInputBand: function (band) 
#>     encode: function () 
#>     index: NULL
#>     initialize: function (xml = NULL) 
#>     inputCoverageBands: list
#>     rootName: coverageBand
#>     setCompositionType: function (compositionType) 
#>     setDefinition: function (definition) 
#>     setIndex: function (index) 
#>     setName: function (name)