A GeoServer access control list service rule

This class models a GeoServer access control list service rule

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer access control list service rule

Details

Geoserver REST API Access Control List Service Rule

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geosapi::GSRESTResource -> geosapi::GSRule -> GSServiceRule

Public fields

roles

one or more roles

Methods

Public methods

GSServiceRule$new()

GSServiceRule$clone()

Inherited methods

geosapi::GSRESTResource$getClassName()

geosapi::GSRESTResource$print()

geosapi::GSRule$decode()

geosapi::GSRule$encode()

Method new()

Initializes a GSLayerRule

Usage

GSServiceRule$new(
  xml = NULL,
  service,
  method,
  permission = c("r", "w", "a"),
  roles
)

Arguments

xml

an object of class xml_node-class

service

service subject to the access control rule, eg. 'wfs'

method

service method subject to the access control rule, eg. 'GetFeature'

permission

the rule permission, either r (read), w (write) or a (administer)

roles

one or more roles to add for the rule

Method clone()

The objects of this class are cloneable with this method.

Usage

GSServiceRule$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.