Geoserver REST API Workspace Setting

Geoserver REST API Workspace Setting

Geoserver REST API Workspace Setting

Format

R6Class object.

Value

Object of R6Class for modelling a GeoServer workspace settings

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geosapi::GSRESTResource -> GSWorkspaceSettings

Public fields

contact

contact

charset

charset

numDecimals

number of decimal

onlineResource

online resource

verbose

verbose

verboseExceptions

verbose exceptions

localWorkspaceIncludesPrefix

local workspace includes prefix

Methods

Public methods

GSWorkspaceSettings$new()

GSWorkspaceSettings$decode()

GSWorkspaceSettings$setCharset()

GSWorkspaceSettings$setNumDecimals()

GSWorkspaceSettings$setOnlineResource()

GSWorkspaceSettings$setVerbose()

GSWorkspaceSettings$setVerboseExceptions()

GSWorkspaceSettings$setLocalWorkspaceIncludesPrefix()

GSWorkspaceSettings$clone()

Inherited methods

geosapi::GSRESTResource$encode()

Method new()

This method is used to instantiate a GSWorkspaceSettings. This settings object is required to activate a workspace configuration, using the method GSManager$createWorkspaceSettings. Supported from GeoServer 2.12

Usage

GSWorkspaceSettings$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method decode()

Decodes from XML

Usage

GSWorkspaceSettings$decode(xml)

Arguments

xml

object of class XMLInternalNode-class

Method setCharset()

Set charset

Usage

GSWorkspaceSettings$setCharset(charset)

Arguments

charset

charset

Method setNumDecimals()

Set number of decimals

Usage

GSWorkspaceSettings$setNumDecimals(numDecimals)

Arguments

numDecimals

number of decimals

Method setOnlineResource()

Set online resource

Usage

GSWorkspaceSettings$setOnlineResource(onlineResource)

Arguments

onlineResource

online resource

Method setVerbose()

Set verbose

Usage

GSWorkspaceSettings$setVerbose(verbose)

Arguments

verbose

verbose

Method setVerboseExceptions()

Set verbose exceptions

Usage

GSWorkspaceSettings$setVerboseExceptions(verboseExceptions)

Arguments

verboseExceptions

verbose exceptions

Method setLocalWorkspaceIncludesPrefix()

Set local workspace includes prefix

Usage

GSWorkspaceSettings$setLocalWorkspaceIncludesPrefix(includesPrefix)

Arguments

includesPrefix

includes prefix

Method clone()

The objects of this class are cloneable with this method.

Usage

GSWorkspaceSettings$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  settings <- GSWorkspaceSettings$new()
  settings$setCharset("UTF-8")
  settings$setNumDecimals(5)